backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
dlib==19.24.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4